Added event sorting, started work on event groups.
Events displayed below the monthly calendar are sorted according to the time they happen in the day. The day displayed below the calendar is updated when the user returns from adding an event. Started work on Groups by adding a menu with the option to either create an event or add a group.